+++ packages/light-apps/src/Accounts/Add/Restore/index.ts
@@ -2,5 +2,5 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-export * from './ImportWithJson';
-export * from './ImportWithPhrase';
+export * from './WithJson';
+export * from './WithPhrase';
